Fix nits
