+++ app/src/components/datasets/DatasetContentTabs/DatasetContentTabs.tsx
@@ -3,9 +3,11 @@ import General from "./General/General";
 import Evaluation from "./Evaluation/Evaluation";
 import Models from "./Models/Models";
 
+export const GENERAL_TAB_KEY = "general";
+
 const tabs = [
   {
-    key: "general",
+    key: GENERAL_TAB_KEY,
     title: "General",
     component: <General />,
   },
