+++ app/src/components/datasets/DatasetContentTabs/Models/Models.tsx
@@ -1,20 +1,32 @@
-import { Card, VStack, HStack, Text } from "@chakra-ui/react";
+import { Card, VStack, HStack, Text, Button } from "@chakra-ui/react";
 import Link from "next/link";
+import { useRouter } from "next/router";
 import ColoredPercent from "~/components/ColoredPercent";
+import { setActiveTab } from "~/components/ContentTabs";
 import { getStatusColor } from "~/components/fineTunes/FineTunesTable";
 import { displayBaseModel } from "~/utils/baseModels";
 import dayjs from "~/utils/dayjs";
 
 import { useDatasetFineTunes } from "~/utils/hooks";
+import { GENERAL_TAB_KEY } from "../DatasetContentTabs";
 
 const Models = () => {
   const fineTunes = useDatasetFineTunes().data;
+
+  const router = useRouter();
 
   if (!fineTunes?.length) {
     return (
       <Text px={8}>
         No models have been trained on this dataset. You can start training a model in the{" "}
-        <b>General</b> tab.
+        <Button
+          variant="link"
+          _hover={{ textDecor: "underline" }}
+          onClick={() => setActiveTab(GENERAL_TAB_KEY, router)}
+        >
+          General
+        </Button>{" "}
+        tab.
       </Text>
     );
   }
